+++ data/graphic/shaders/scalable_kinect_fusion/vs_rasterize_grid.glsl
@@ -4,13 +4,13 @@
 
 #include "scalable_kinect_fusion/common_tracking.glsl"
 
-uvec3 Indexto3D(uint Index, uint Resolution)
+vec3 Indexto3D(int Index, int Resolution)
 {
-    uint z = Index / (Resolution * Resolution);
+    int z = Index / (Resolution * Resolution);
     Index -= (z * Resolution * Resolution);
-    uint y = Index / Resolution;
-    uint x = Index % Resolution;
-    return uvec3(x, y, z);
+    int y = Index / Resolution;
+    int x = Index % Resolution;
+    return vec3(x, y, z);
 }
 
 // -----------------------------------------------------------------------------
@@ -19,7 +19,7 @@ uvec3 Indexto3D(uint Index, uint Resolution)
 
 layout(row_major, std140, binding = 2) uniform UBOTransform
 {
-    uint g_Resolution;
+    int g_Resolution;
     float g_CubeSize;
     float g_ParentSize;
     ivec3 g_Offset;
